+++ ContentPatcher/release-notes.md
@@ -3,6 +3,7 @@
 # Release notes
 ## Upcoming release
 * Fixed `Day` token not allowing zero values.
+* Fixed dynamic tokens validated before they're ready.
 
 ## 1.9.1
 Released 12 June 2019.