+++ ContentPatcher/release-notes.md
@@ -1,7 +1,9 @@
 [← back to readme](README.md)
 
 # Release notes
-## Upcoming release
+## 1.9.2
+Released 25 July 2019.
+
 * Fixed `Day` token not allowing zero values.
 * Fixed dynamic tokens validated before they're ready.
 * Fixed custom tokens called with non-ready inputs in some cases.